Load ClusterStateListenerModule by default
