+++ docs/operation.md
@@ -45,7 +45,6 @@ to config file's modules section like below
 ```yaml
 modules:
   - io.trino.gateway.ha.module.HaGatewayProviderModule
-  - io.trino.gateway.ha.module.ClusterStateListenerModule
   - io.trino.gateway.ha.module.QueryCountBasedRouterProvider
 ```
 - The router works on the stats it receives from the clusters about the load i.e number queries queued and running on a cluster at regular intervals which can be configured like below. The default interval is 1 min
