+++ gateway-ha/gateway-ha-config-docker.yml
@@ -18,7 +18,6 @@ clusterStatsConfiguration:
 
 modules:
   - io.trino.gateway.ha.module.HaGatewayProviderModule
-  - io.trino.gateway.ha.module.ClusterStateListenerModule
 
 managedApps:
   - io.trino.gateway.ha.clustermonitor.ActiveClusterMonitor
