+++ gateway-ha/src/test/resources/test-config-with-routing-template.yml
@@ -11,7 +11,6 @@ dataStore:
 
 modules:
   - io.trino.gateway.ha.module.HaGatewayProviderModule
-  - io.trino.gateway.ha.module.ClusterStateListenerModule
 
 managedApps:
   - io.trino.gateway.ha.clustermonitor.ActiveClusterMonitor
